requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

bcoding==1.5
    # via -r requirements.in
decorator==4.4.2
    # via
    #   retry
    #   validators
dnspython==2.0.0
    # via ipwhois
future==0.18.2
    # via python-whois
ipwhois==1.2.0
    # via -r requirements.in
publicsuffix2==2.20191221
    # via -r requirements.in
py==1.10.0
    # via retry
python-whois==0.7.3
    # via -r requirements.in
retry==0.9.2
    # via -r requirements.in
six==1.15.0
    # via validators
validators==0.18.2
    # via -r requirements.in
